:truck: Moved backend functions into helpers
